feature: sync node taint to pod condition (#263)
For pre-activities, our users might only care about specific pod annotations.
In our current setup, they would have to watch for node updates, check if the taint was added, then figure out if the node is hosting a pod they care about, and finally trigger the update mechanism.
This makes the whole process much more complex than expected.

Therefore, we've decided to create a new condition on all pods that have pre-activities assigned. This condition should always be in sync with the NLA node taint, so that the operators can react properly.