Add id to conditions, use as string rep. (#269)
* Add id to conditions, use as string rep.

The key in `--node-condition` will now represent the human readable id for a
condition. Doing this to allow a human readable id when supplying a
`--node-condition` would not have a meaningful id from the type alone. For example,
condition with type=Ready, conditionStatus=Unknown would simply show up
as `ready` without this change.

By default, if the `type` is not specified the `--node-condition` key
will by used as the `type` preserving the old behavior.

continuation of https://github.com/DataDog/draino/pull/268

* cr